requirements.txt
# Web框架
fastapi==0.95.1
uvicorn[standard]==0.22.0
websockets==11.0.2
httpx==0.24.0

# 数据处理和机器学习
numpy==1.24.3
pandas==2.0.1
scikit-learn==1.2.2
qlib>=0.9.1
lightgbm==3.3.5
torch==2.0.1

# 加密货币交易
ccxt==3.0.74
python-binance==1.0.17

# 数据存储
motor==3.1.2
pymongo==4.3.3
redis==4.5.5
aioredis==2.0.1

# 工具库
python-dotenv==1.0.0
pydantic==1.10.7
eth-hash[pycryptodome]==0.5.1
requests==2.30.0
pytz==2023.3
ujson==5.7.0
aiohttp==3.8.4

# 日志和监控
prometheus-client==0.16.0
structlog==23.1.0

# 开发工具
pytest==7.3.1
pytest-asyncio==0.21.0
black==23.3.0
isort==5.12.0
mypy==1.3.0 